# packages/api-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APMTRACK API",
    "version": "0.1.0",
    "description": "Contrat OpenAPI backend APMTRACK: sante, auth, RBAC, communes, utilisateurs, agents, zones, referentiel, PV, paiements, signalements, patrouilles, dashboard, audit logs et exports."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Developpement local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/dashboard/summary": {
      "get": {
        "summary": "Dashboard summary",
        "responses": {
          "200": {
            "description": "Succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/dashboard/pvs": {
      "get": {
        "summary": "Dashboard pvs",
        "responses": {
          "200": {
            "description": "Succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/dashboard/payments": {
      "get": {
        "summary": "Dashboard payments",
        "responses": {
          "200": {
            "description": "Succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/dashboard/agents": {
      "get": {
        "summary": "Dashboard agents",
        "responses": {
          "200": {
            "description": "Succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/dashboard/signalements": {
      "get": {
        "summary": "Dashboard signalements",
        "responses": {
          "200": {
            "description": "Succes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObject"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentification invalide",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Role ou commune non autorise",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      },
...
      "GenericObject": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  }
}